chore: removed 'dashboard' service
