+++ packages/common/src/services/index.ts
@@ -2,7 +2,6 @@ export { AddressService } from "./address.service";
 export { CategoryService } from "./category.service";
 export { CollectionService } from "./collection.service";
 export { CustomerService } from "./customer.service";
-export { DashboardService } from "./dashboard.service";
 export { LaunchNotificationService } from "./launch-notification.service";
 export { NotificationService } from "./notification.service";
 export {
