+++ packages/common/src/services/store.service.ts
@@ -12,9 +12,19 @@ import {
   enhanceLogWithTraceContext,
   traceStaticClass,
 } from "@dukkani/tracing";
+import {
+  endOfDay,
+  endOfWeek,
+  startOfDay,
+  startOfWeek,
+  subDays,
+} from "date-fns";
+import { OrderQuery } from "../entities/order/query";
+import { OrderItemQuery } from "../entities/order-item/query";
 import { ProductQuery } from "../entities/product/query";
 import { StoreEntity } from "../entities/store/entity";
 import { StoreQuery } from "../entities/store/query";
+import { OrderStatus } from "../schemas/order/enums";
 import { isReservedStoreSlug } from "../schemas/store/constants";
 import type { CreateStoreOnboardingInput } from "../schemas/store/input";
 import type {
@@ -335,6 +345,114 @@ class StoreServiceBase {
     });
 
     return StoreEntity.getSimpleRo(updatedStore);
+  }
+
+  /**
+   * Dashboard home overview: today / week order and revenue stats for the user's store(s).
+   * When `storeId` is set, stats are limited to that store; otherwise all owned stores.
+   */
+  static async getStoreStats(userId: string, storeId?: string) {
+    addSpanAttributes({
+      "store.stats_user_id": userId,
+      "store.stats_filter_id": storeId ?? "all",
+    });
+
+    const userStoreIds = await database.store.findMany({
+      where: { ownerId: userId },
+      select: { id: true },
+    });
+
+    const storeIds = userStoreIds.map((s) => s.id);
+
+    if (storeIds.length === 0) {
+      return {
+        todayOrders: 0,
+        todayOrdersChange: 0,
+        todayRevenue: 0,
+        weekOrders: 0,
+      };
+    }
+
+    const filteredStoreIds =
+      storeId && storeIds.includes(storeId) ? [storeId] : storeIds;
+
+    const now = new Date();
+    const todayStart = startOfDay(now);
+    const todayEnd = endOfDay(now);
+    const yesterdayStart = startOfDay(subDays(now, 1));
+    const yesterdayEnd = endOfDay(subDays(now, 1));
+    const weekStart = startOfWeek(now, { weekStartsOn: 1 });
+    const weekEnd = endOfWeek(now, { weekStartsOn: 1 });
+
+    const baseOrderWhere = OrderQuery.getWhere(filteredStoreIds);
+
+    const [todayOrders, yesterdayOrders, todayDeliveredOrders, weekOrders] =
+      await database.$transaction([
+        database.order.count({
+          where: {
+            ...baseOrderWhere,
+            createdAt: {
+              gte: todayStart,
+              lte: todayEnd,
+            },
+          },
+        }),
+        database.order.count({
+          where: {
+            ...baseOrderWhere,
+            createdAt: {
+              gte: yesterdayStart,
+              lte: yesterdayEnd,
+            },
+          },
+        }),
+        database.order.findMany({
+          where: {
+            ...baseOrderWhere,
+            status: OrderStatus.DELIVERED,
+            createdAt: {
+              gte: todayStart,
+              lte: todayEnd,
+            },
+          },
+          include: {
+            orderItems: {
+              select: OrderItemQuery.getRevenueSelect(),
+            },
+          },
+        }),
+        database.order.count({
+          where: {
+            ...baseOrderWhere,
+            createdAt: {
+              gte: weekStart,
+              lte: weekEnd,
+            },
+          },
+        }),
+      ]);
+
+    const todayRevenue = todayDeliveredOrders.reduce((sum, order) => {
+      const orderTotal = order.orderItems.reduce(
+        (itemSum, item) => itemSum + Number(item.price) * item.quantity,
+        0,
+      );
+      return sum + orderTotal;
+    }, 0);
+
+    const todayOrdersChange =
+      yesterdayOrders > 0 ? todayOrders - yesterdayOrders : todayOrders;
+
+    addSpanAttributes({
+      "store.stats_owner_store_count": storeIds.length,
+    });
+
+    return {
+      todayOrders,
+      todayOrdersChange,
+      todayRevenue,
+      weekOrders,
+    };
   }
 }
